requirements.txt
requests
feedparser
beautifulsoup4
jinja2
pyyaml
python-dateutil
